Tj/dono amount (#11)
* lfg

* lowered the min donation on the frontend

* changed int to float to allow for sub dollar donos

* did some AI filtering

---------

+++ src/app/api/minimum/route.ts
@@ -4,7 +4,7 @@ import { NextResponse } from 'next/server';
 export async function GET(request: Request) {
 
     const requestUrl = new URL(request.url);
-    const minimum = parseInt(requestUrl.searchParams.get("minimum")!)
+    const minimum = parseFloat(requestUrl.searchParams.get("minimum")!)
     const pubkey = requestUrl.searchParams.get("pubkey")!;
     
     await updateMinimum(minimum, pubkey)
